added cpp and h file, added to mbuild
