+++ Utilities/CSV.h
@@ -1,7 +1,6 @@
 
 #pragma once
 
-#include "Utilities.h"
 
 #include <algorithm>
 #include <cmath>
@@ -20,10 +19,20 @@
 // csv lines can also be directly converted to a vector of a
 // templated type. Note: Behaviour is uspecified if all values
 // in the csv string cannot be converted to the templated type
+
 class CSVReader {
 
   std::string delimiter_, quotation_;
   std::regex item_;
+  
+  // try and convert a std::string to a particular type
+  // warning: no error if value is not  valid type
+  template <typename T> inline static auto stringTo(std::string source) {
+    std::stringstream ss(source);
+    T target;
+    ss >> target;
+    return target;
+  }
 
 public:
   CSVReader() : CSVReader(',', '"') {}
@@ -31,43 +40,58 @@ public:
       : delimiter_(1, s), quotation_(1, se),
         item_(R"((.*?)()" + delimiter_ + "|" + quotation_ + R"(|$))") {}
 
+ // template <class T> auto  stringTo(std::string source);
   // parse a csv line into a vector<T>
   template <typename T = std::string>
-  auto parseLine(std::string &raw_line) const;
-};
+  std::vector<T> parseLine(std::string raw_line) {
+    std::vector<T> data;
 
-template <typename T> auto CSVReader::parseLine(std::string &raw_line) const {
-  std::vector<T> data;
-  auto current = 0u;
+	if (raw_line.empty())
+		return data;
+/*
+    static const std::regex stripoff_qoute(R"(^"(.*?)?"$)");
+    static const std::regex stripoff_square_brackets(R"(^\[(.*?)\]$)");
+    std::smatch m_quote;
+    raw_line = std::regex_match(raw_line, m_quote, stripoff_qoute)
+                   ? m_quote[1].str()
+                   : raw_line;
+    std::smatch m_square;
+    raw_line = std::regex_match(raw_line, m_square, stripoff_square_brackets)
+                   ? m_square[1].str()
+                   : raw_line;
+*/
 
-  while (true) {
-    // find next delimiter
-    auto delim = raw_line.find_first_of(delimiter_, current);
-    // find next quotation
-    auto quote = raw_line.find_first_of(quotation_, current);
-    // if the next quotation comes before the next delimiter
-    if (quote < delim) {
-      //  find next quotation
-      auto nested_quote = raw_line.find_first_of(quotation_, quote + 1);
-      // find first delimiter after that; ignoring delimiter inside of quotation
-      // warning: assumes that quotes come in pairs
-   	  delim = raw_line.find_first_of(delimiter_, nested_quote);
+    auto current = 0u;
+
+    while (true) {
+      // find next delimiter
+      auto delim = raw_line.find_first_of(delimiter_, current);
+      // find next quotation
+      auto quote = raw_line.find_first_of(quotation_, current);
+      // if the next quotation comes before the next delimiter
+      if (quote < delim) {
+        //  find next quotation
+        auto nested_quote = raw_line.find_first_of(quotation_, quote + 1);
+        // find first delimiter after that; ignoring delimiter inside of
+        // quotation warning: assumes that quotes come in pairs
+        delim = raw_line.find_first_of(delimiter_, nested_quote);
+      }
+      // if line is completely parsed
+      if (delim == std::string::npos) {
+        break;
+      }
+      // add the next field
+      data.push_back(stringTo<T>(raw_line.substr(current, delim - current)));
+      // search for the next field
+      current = delim + 1;
     }
-    // if line is completely parsed
-    if (delim == std::string::npos) {
-      break;
-    }
-	// add the next field
-    data.push_back(stringTo<T>(raw_line.substr(current, delim - current)));
-    // search for the next field
-	current = delim + 1;
+
+    // add the last field
+    data.push_back(stringTo<T>(raw_line.substr(current)));
+
+    return data;
   }
-  
-  //add the last field
-  data.push_back(stringTo<T>(raw_line.substr(current)));
-
-  return data;
-}
+};
 
 // parses a csv file and stores in memory.
 // the first line of the file is treated as the column headers
@@ -88,31 +112,34 @@ public:
   CSV(std::string fn, char s, char se);
   CSV(std::string fn) : CSV(fn, ',', '"') {}
 
-  // look up a value in a colmn and return the value in the corresponding row of
-  // the other column
-  auto lookUp(std::string lookup_column, std::string value,
-              std::string return_column) const;
+  // return csv file name
+  auto fileName() const { return file_name_; }
+
+  // number of columns in the file
+  auto column_count() const { return columns_.size(); }
 
   // number of rows in the file
   auto row_count() const { return rows_.size(); }
-
-  // number of columns in the file
+  
+  // return all columns in the file
   auto columns() const { return columns_; }
-
-  // return all values corresponding to a single column
-  auto singleColumn (std::string column);
 
   // return all rows in the file
   auto rows() const { return rows_; }
 
-  // merges another csv file. Only extra columns are updated. All values in the
+  // return all values corresponding to a single column
+  std::vector<std::string> singleColumn(std::string column);
+
+  // look up a value in a column and return the value in the corresponding row of
+  // the other column
+  std::string lookUp(std::string lookup_column, std::string value,
+                     std::string return_column) const;
+
+  // merges another csv file. Only extra columns are added. All values in the
   // specified column must also exist in the same column in the other file.
   // only rows in the other file that have matching values in this
   // file are merged
   void merge(CSV csv, std::string column);
-
-  // return csv file name
-  auto fileName() const { return file_name_; }
 
   // check existence of a column
   bool hasColumn(std::string name) const {
@@ -120,171 +147,4 @@ public:
            std::end(columns_);
   }
 };
-
-
-auto CSV::singleColumn(std::string column) {
-  if (!hasColumn(column)) {
-    std::cout << " Error : could not find column " << column
-              << " to merge from file " << file_name_ << std::endl;
-    exit(1);
-  }
-
-  auto const column_pos =
-      std::find(std::begin(columns_), std::end(columns_), column) -
-      std::begin(columns_);
-
-  std::vector<std::string> values;
-  std::transform(std::begin(rows_), std::end(rows_), std::back_inserter(values),
-                 [column_pos](auto const &row) { return row[column_pos]; });
-
-  return values;
-}
-
-auto CSV::lookUp(std::string lookup_column, std::string value,
-                 std::string return_column) const {
-
-  if (!hasColumn(lookup_column)) {
-    std::cout << " Error : could not find column " << lookup_column
-              << " in file " << file_name_ << std::endl;
-    exit(1);
-  }
-
-  if (!hasColumn(return_column)) {
-    std::cout << " Error : could not find column " << return_column
-              << " in file " << file_name_ << std::endl;
-    exit(1);
-  }
-
-  // find position of lookup column
-  auto const column_pos =
-      std::find(std::begin(columns_), std::end(columns_), lookup_column) -
-      std::begin(columns_);
-
-  // find number of values in lookup column
-  auto const value_count = std::count_if(
-      std::begin(rows_), std::end(rows_),
-      [value, column_pos](auto &row) { return row[column_pos] == value; });
-
-  if (!value_count) {
-    std::cout << "Error : could not find requested lookup value" << value
-              << " from column " << lookup_column << " from file " << file_name_
-              << std::endl;
-    exit(1);
-  }
-
-  if (value_count > 1) {
-    std::cout << "Error : multiple entries found for requested lookup value"
-              << value << " from column " << lookup_column << " from file "
-              << file_name_ << std::endl;
-    exit(1);
-  }
-
-  // find row where lookup column has value
-  auto const value_pos = std::find_if(std::begin(rows_), std::end(rows_),
-                                      [value, column_pos](auto &row) {
-                                        return row[column_pos] == value;
-                                      }) -
-                         std::begin(rows_);
-
-  // find position of return column
-  auto const return_pos =
-      std::find(std::begin(columns_), std::end(columns_), return_column) -
-      std::begin(columns_);
-
-  return rows_[value_pos][return_pos];
-}
-
-CSV::CSV(std::string fn, char s, char se) : file_name_(fn), reader_(s, se) {
-
-  std::ifstream file(file_name_);
-  if (!file.is_open()) {
-    std::cout << " Error: readColumnsFromCSVFile cannot open file "
-              << file_name_ << std::endl;
-    exit(1);
-  }
-
-  std::string raw_line;
-
-  // read header line
-  getline(file, raw_line);
-  columns_ = reader_.parseLine(raw_line);
-
-  // ensure column names are unique
-  std::set<std::string> uniq_headers(std::begin(columns_), std::end(columns_));
-  if (uniq_headers.size() != columns_.size()) {
-    std::cout << "Error: CSV file " << file_name_
-              << " does not have unique Header names" << std::endl;
-    exit(1);
-  }
-
-  // read remaining rows
-  while (getline(file, raw_line)) {
-    auto data_line = reader_.parseLine(raw_line);
-    // ensure all rows have correct number of columns
-    if (columns_.size() != data_line.size()) {
-      std::cout << " Error: incorrect number of columns in CSV file "
-                << file_name_ << std::endl;
-      exit(1);
-    }
-    rows_.push_back(data_line);
-  }
-}
-
-void CSV::merge(CSV merge_csv, std::string column) {
-
-  if (!hasColumn(column)) {
-    std::cout << " Error : could not find column " << column << " in file "
-              << file_name_ << std::endl;
-    exit(1);
-  }
-
-  if (!merge_csv.hasColumn(column)) {
-    std::cout << " Error : could not find column " << column
-              << " to merge from file " << merge_csv.fileName() << std::endl;
-    exit(1);
-  }
-
-  // ensure lookup column has distinct values
-  auto lookop_column = singleColumn(column);
-  std::set<std::string> lookup_values(std::begin(lookop_column),
-                                      std::end(lookop_column));
-  if (lookup_values.size() != rows_.size()) {
-    std::cout << "Error: CSV file " << file_name_
-              << " does not have unique values in column " << column
-              << std::endl;
-    exit(1);
-  }
-
-  // ensure column in second file has matching values for all values in this
-  // file
-  auto merge_column = merge_csv.singleColumn(column);
-  std::set<std::string> merge_values(std::begin(merge_column),
-                                     std::end(merge_column));
-  if (!std::includes(std::begin(merge_values), std::end(merge_values),
-                     std::begin(lookup_values), std::end(lookup_values))) {
-    std::cout << "Error: CSV file " << merge_csv.fileName()
-              << " does not have some matching values for column " << column
-              << std::endl;
-    exit(1);
-  }
-
-  // find position of lookup column
-  auto const column_pos =
-      std::find(std::begin(columns_), std::end(columns_), column) -
-      std::begin(columns_);
-
-  // merge columns from second file
-  for (auto const &merge_column : merge_csv.columns()) {
-    // only add additional columns
-    if (std::find(std::begin(columns_), std::end(columns_), merge_column) ==
-        std::end(columns_)) {
-      columns_.push_back(merge_column);
-      // for each column add value to every row
-      for (auto &row : rows_)
-        row.push_back(merge_csv.lookUp(column, row[column_pos], merge_column));
-    }
-  }
-}
-
-
